세부 수정

+++ src/views/ReviewPage/MyReview.js
@@ -5,12 +5,15 @@ import "./ReviewHome.css";
 import Header from "../Header/Header";
 import axios from "axios";
 import useDidMountEffect from "./useDidMountEffect";
+import { ToggleButton } from "@mui/material";
 
 function MyReview() {
   const acToken = sessionStorage.getItem("accesstoken");
 
   const [reviews, setReviews] = useState([]);
   let hallId = "";
+
+  const [toggle, setToggle] = useState(true);
 
   const myUrl = "/api/v1/reviews";
 
@@ -33,10 +36,9 @@ function MyReview() {
           response.data.result !== "undefined" &&
           response.data.result !== null
         )
-          //setData(response.data.result);
           setReviews(response.data.result);
-        console.log(reviews);
-        //setToggle(false);
+        console.log(response.data.result);
+        setToggle(true);
       })
       .catch((error) => {
         console.log(error.response.data.errorMessage);
@@ -67,14 +69,16 @@ function MyReview() {
 
   //삭제
   const reivewDelete = (reId, reName) => {
-    deleteUrl = deleteUrl + reName + "&reviewId=" + reId;
-    console.log(deleteUrl);
-    deleteReview();
-    //setToggle(true);
+    if (window.confirm("리뷰를 삭제하시겠습니까?")) {
+      deleteUrl = deleteUrl + reName + "&reviewId=" + reId;
+      console.log(deleteUrl);
+      deleteReview();
+    }
   };
+
   return (
     <div>
-      {reviews.length > 0 ? (
+      {toggle > 0 ? (
         <div>
           {reviews.map((review) => (
             <div key={review.reviewId}>
@@ -90,9 +94,10 @@ function MyReview() {
                 number={review.number}
                 starPoint={review.starPoint}
                 reviewId={review.reviewId}
+                createdAt={review.createdAt}
               />
               <div className="reviewED">
-                <button>수정</button>
+                {/* <button>수정</button> */}
                 <button
                   onClick={() => reivewDelete(review.reviewId, review.hallName)}
                 >
